Fix other issues with readonly caches, refs #10906

+++ src/Composer/Cache.php
@@ -104,7 +104,7 @@ class Cache
                     || !is_writable($this->root)
                 )
             ) {
-                $this->io->writeError('<warning>Cannot create cache directory ' . $this->root . ', or directory is not writable. Proceeding without cache</warning>');
+                $this->io->writeError('<warning>Cannot create cache directory ' . $this->root . ', or directory is not writable. Proceeding without cache. See also cache-read-only config if your filesystem is read-only.</warning>');
                 $this->enabled = false;
             }
         }
@@ -265,7 +265,7 @@ class Cache
      */
     public function remove($file)
     {
-        if ($this->isEnabled()) {
+        if ($this->isEnabled() && !$this->readOnly) {
             $file = Preg::replace('{[^'.$this->allowlist.']}i', '-', $file);
             if (file_exists($this->root . $file)) {
                 return $this->filesystem->unlink($this->root . $file);
@@ -280,7 +280,7 @@ class Cache
      */
     public function clear()
     {
-        if ($this->isEnabled()) {
+        if ($this->isEnabled() && !$this->readOnly) {
             $this->filesystem->emptyDirectory($this->root);
 
             return true;
@@ -314,7 +314,7 @@ class Cache
      */
     public function gc($ttl, $maxSize)
     {
-        if ($this->isEnabled()) {
+        if ($this->isEnabled() && !$this->readOnly) {
             $expire = new \DateTime();
             $expire->modify('-'.$ttl.' seconds');
